grantsetu-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
celery[redis]==5.4.0
redis==5.2.1
pydantic-settings==2.7.1
pydantic[email]==2.10.5
httpx==0.28.1
aiohttp==3.11.11
beautifulsoup4==4.12.3
lxml==5.3.0
python-slugify==8.0.4
PyMuPDF==1.25.3
resend==2.7.0
structlog==24.4.0
python-multipart==0.0.20
PyJWT==2.9.0
